+++ server/routes/spa.js
@@ -1,0 +1,12 @@
+import path from 'path';
+
+export default {
+    method: 'GET',
+    path: '/{param*}',
+    handler: {
+      directory: {
+        path: path.resolve('/app/dist'),
+        index: ['index.html'],
+      },
+    },
+};
